Restructure actionsTypes
