+++ app/constants/actionTypes.js
